Add user validations
There is no presence validation for the password because records
retrieved from the database won't have a password, just a password hash.
The custom email validation contains a minimal regex to make sure the
email is formatted appropriately, but a mailer to check that the email
is an existing address should be implemented in the future.