Fixe a number of users controller tests
The tests were specifying the format of the response in the requests.
This was necessary in tests involving views so the correct format view
could be found in the absence of the router, but it made the tests
redundant in the case of controller actions that were meant to return
plain json.