Set up minimal provider container with store
This container will wrap the entire app, giving the components access to
the store. The store itself was configured with a thunk middleware so
that async action creators can be managed and has the option of adding a
preloaded state, which we'll use so that session info can be loaded
between visits when the user hasn't logged out.