Fix the createNewUser tests
The previous implementation of the postUser tests weren't truly unit
tests because they failed to mock the api functions the function being
tested relied on. Instead they mocked the ajax call. The change to
postUser broke the previous test since the token couldn't be stubbed in.
This new implementation mocks the api function directly to fix it.