Make it so the current user persists across reload
Previously, there was no way for the frontend to know whether there was
an ongoing session, which meant that reloading the page always yielded
an unauthenticated version. By bootstrapping current_user to the window,
we can incorporate the current user into our store's initial state to
allow sessions to persist across reloads.