Add logging out functionality
Because logging out required the user be redirected towards the log in
page, the logging out functionality required the implementation of
protected and authentication routes to let the frontend manage those
redirects without page reload.

Previously, any page would be accessible regardless of whether the user
was logged in. But now, attempting to redirect to the access the
homepage without being logged in will redirect the user to log in (or
sign up) first.