Fix the tests to use a helper for examples
Previously, examples were generated by creating a new DateTime object
directly in the test block. This made it hard to parse out what the date
in question was and why the the result of the test ought to be what it
is. Using the helpers to set start and end dates makes the tests shorter
and easier to understand.