Change all recurring items specs to use helper
The helpers that were previously in the MonthsItem spec is now located in
its own helper file so that setting up examples for recurring type
budget items can be a little bit more streamlined. This changed allowed
us to clean up the tests for DaysUnitItem so they're easier to read.