Fix routes so items aren't nested within users
Since only one user is going to be logged at a time and all items are
going to be created for that particular user, there is no need to have
the create and index routes for recurring and non-recurring items nested
under user. There is no ambiguity as to which user they should belong
to.