Fix budget items controller
Previously, the contollers would need a user parameter to determine who
the user the item in question belonged to, but the routes have been
adjusted so that no such parameter exists, so the the controllers needed
to be modified to use the current_user instead. This change makes sense,
as a user can only interact with the items they own and it simplifies
the api calls.