Fix the index action on the items controllers
The index action hadn't been fixed to reflect using the current_user
instead of a user parameter defined in the request path, which meant
that using the API to request all of the recurring or non-recurring
items would return an forbidden error. This has now been fixed.

The fix also necessitated tweaking the ensure_owner_signed_in! function
of the controllers. Since checking that a user is logged is enough to
check that the owner of the items is logged in, the check could be
removed for the index action. The check could therefore also be
simplified, since all remaining actions come with an associated id
parameter in the query string.