Add the entities reducer
This reducer combines two reducers that handle recurring items and
non-recurring items separately, and will allow us to manage the
part of the app state that deals with creating, displaying and deleting
items. Much of the logic of the items reducers was the same, so it got
abstracted into a itemsReducer function that generates the particular
reducers given the relevant actions.