Fix ensure_owner_signed_in! in item controllers
There were a few instances where the owner of the item being requested
could not be established because the controller had been handed an
invalid id. This would cause the method to fail with an error. The
method was fixed to handle this gracefully by simply setting the owner
to nil. Because the logic was shared across both item controllers, a new
BudgetItemsContoller superclass was created to hold the logic.