Add selectors for ordering items by day
The budget display page will need to be able to display each of the
items that have and will be incurred that month. This selector returns
an object that uses the month date's as keys, holding an array with
each of the relevant items, to make that task a little easier.