Add a component to display the month's items
The MonthlyBudget component displays a list of all the income items and expenses
that the user has incurred during the month, both recurring and
non-recurring. This allows them to get, at a glance, a sense of where
their money has is going and where most of their money that month will
be coming from.