Add error rendering to the user forms
The user forms now handle all the form errors, by updating the state
when the server responds with an error array and rendering all the
errors if present. It will also clear the form once a successful request
has been received or when the component is mounted again, to to prevent
errors persisting when it's not appropriate.