Fix ItemList's display logic
Previously, ItemList would utilise its own internal state to determine
which kinds of items it should display. While this worked fine for most
cases, it was problematic when new items were added, as there was no way
to alter which kind of item was being displayed from the form component,
so it wouldn't be immediately obvious to the user that an item had been
added if the wrong kind of item was being displayed. Making the list's
display logic rely on application-wide state makes this easy to fix.